+++ lib/presentation/blocs/category/category_bloc.dart
@@ -1,4 +1,8 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import '../../../core/constants/default_categories.dart';
+import '../../../domain/entities/category.dart' as entities;
 import '../../../domain/repositories/category_repository.dart';
 import 'category_event.dart';
 import 'category_state.dart';
@@ -18,7 +22,23 @@ class CategoryBloc extends Bloc<CategoryEvent, CategoryState> {
   Future<void> _onLoad(CategoryLoadRequested event, Emitter<CategoryState> emit) async {
     emit(const CategoryLoading());
     try {
-      final categories = await _repository.fetchAll();
+      var categories = await _repository.fetchAll();
+      if (categories.isEmpty) {
+        final prefs = await SharedPreferences.getInstance();
+        final seeded = prefs.getBool('categories_seeded_v1') ?? false;
+        if (!seeded) {
+          for (final entry in DefaultCategories.list) {
+            await _repository.create(entities.Category(
+              name: entry.name,
+              type: entry.type,
+              icon: entry.icon,
+            ));
+          }
+          await prefs.setBool('categories_seeded_v1', true);
+          categories = await _repository.fetchAll();
+          debugPrint('[CategoryBloc] Seeded ${DefaultCategories.list.length} default categories');
+        }
+      }
       emit(CategoryLoaded(categories: categories));
     } catch (e) {
       emit(CategoryError(message: e.toString()));